added ts support, codemirrorAutocompletionExtension logic moved to separate file
