+++ frontend/package.json
@@ -25,6 +25,7 @@
     "eslint": "^8.45.0",
     "eslint-plugin-vue": "^9.15.1",
     "prettier": "^3.0.0",
+    "typescript": "^5.2.2",
     "vite": "^4.4.6"
   }
 }
